pyproject.toml
[project]
name = "f1-race-intelligence-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "fastf1>=3.8.3",
    "langchain>=1.2.16",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-huggingface>=1.2.2",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.10",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
